Rename Translate HOC to translate
