+++ src/i18n/index.js
@@ -3,5 +3,5 @@ export const DEFAULT_LOCALE = 'en';
 export * from './TranslationUtils';
 
 export englishMessages from './messages';
-export Translate from './Translate';
+export translate from './translate';
 export TranslationProvider from './TranslationProvider';
