+++ src/mui/button/RefreshButton.js
@@ -1,7 +1,7 @@
 import React, { PropTypes } from 'react';
 import FlatButton from 'material-ui/FlatButton';
 import NavigationRefresh from 'material-ui/svg-icons/navigation/refresh';
-import Translate from '../../i18n/Translate';
+import translate from '../../i18n/translate';
 
 const RefreshButton = ({ label = 'aor.action.refresh', translate, refresh }) => <FlatButton
     primary
@@ -16,4 +16,4 @@ RefreshButton.propTypes = {
     translate: PropTypes.func.isRequired,
 };
 
-export default Translate(RefreshButton);
+export default translate(RefreshButton);
