+++ src/mui/layout/Notification.js
@@ -2,7 +2,7 @@ import React, { PropTypes } from 'react';
 import { connect } from 'react-redux';
 import Snackbar from 'material-ui/Snackbar';
 import { hideNotification as hideNotificationAction } from '../../actions/notificationActions' ;
-import Translate from '../../i18n/Translate';
+import translate from '../../i18n/translate';
 
 function getStyles(context) {
     if (!context) return { primary1Color: '#00bcd4', accent1Color: '#ff4081' };
@@ -64,7 +64,7 @@ const mapStateToProps = state => ({
     type: state.admin.notification.type,
 });
 
-export default Translate(connect(
+export default translate(connect(
   mapStateToProps,
   { hideNotification: hideNotificationAction },
 )(Notification));
